+++ src/replaceDeep.ts
@@ -1,8 +1,26 @@
+import { replaceDeepByFn } from "./replaceDeepByFn.js";
+
+type Options = {
+    /**
+     * If true, the source objects and arrays will be mutated. Default is false.
+     */
+    mutate?: boolean;
+    /**
+     * If true, the function will go into instances for replacement. Otherwise, it will only replace properties of plain
+     * objects. Default is false.
+     * Warning: This option requires `mutate` to be enabled, because we can't clone instances.
+     */
+    replaceInstancesProps?: boolean;
+};
+
 /**
  * Replaces all occurrences of `search` with `value` in `source` object/array. Comparison is done with `Object.is`.
  * If `source` is exactly the `search` a `value` will be returned. It does not do a substring replacements.
  *
- * It mutates the `source` object/array!
+ * Warnings:
+ * - By default, it does not mutate the `source`/deep objects/arrays, but it can be enabled with `mutate` option.
+ * - By default, it does not go into instances for replacement, only plain objects.
+ * - If your instances are cross-referenced, you may end up in an infinite loop.
  *
  * TypeScript users: This is way too dynamic to type properly, therefore, typing assumes the most basic form of
  * replacement where search and value are of the same type. If that's not the case for you - you'll have to typecast.
@@ -10,33 +28,10 @@
  * @param source - source object/array/value
  * @param search - value to search for
  * @param value - value to replace with
+ * @param options - optional options
  */
-const replaceDeep = <T>(source: T, search: unknown, value: unknown): T => {
-    if (Object.is(source, search)) {
-        return value as T;
-    }
-
-    if (source == null) {
-        return source;
-    }
-
-    if (typeof source === "object") {
-        if (Array.isArray(source)) {
-            for (let i = 0; i < source.length; i++) {
-                // eslint-disable-next-line no-param-reassign,@typescript-eslint/no-unsafe-assignment
-                source[i] = replaceDeep(source[i], search, value);
-            }
-            return source;
-        }
-
-        return Object.keys(source).reduce<Record<string, unknown>>((acc, key) => {
-            // eslint-disable-next-line no-param-reassign
-            acc[key] = replaceDeep((source as Record<string, unknown>)[key], search, value);
-            return acc;
-        }, {}) as T;
-    }
-
-    return source;
+const replaceDeep = <T>(source: T, search: unknown, value: unknown, options?: Options): T => {
+    return replaceDeepByFn(source, (v) => Object.is(v, search), () => value, options);
 };
 
 export {