+++ src/utils/utils.ts
@@ -1,0 +1,11 @@
+class DataWrapper<T> {
+    public data: T;
+
+    public constructor(data: T) {
+        this.data = data;
+    }
+}
+
+export {
+    DataWrapper,
+};