+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The format is based on [EZEZ Changelog](https://ezez.dev/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Breaking
+- `replaceDeep` will not mutate anything by default — previously it was wrongly documented as always mutating,
+but it was not mutating objects. Mutating behavior can be enabled and additionally controlled by defining if instances
+properties are allowed to be mutated.
+- `serialize` to avoid excessive calls to serializer functions - now they are only called with non-plain objects or unknown data types, this breaks some rare use cases
+### Changed
+- `serialize` custom serializers now allow things like `Date` (that defines .toJSON) to be supported as the user would expect
+### Added
+- `replaceDeepByFn` method for more granular control over replacing values. `replaceDeep` is a simplified wrapper over
+the new function
 ### Dev
 - upgraded some jsdocs